requirements.txt
dj-database-url==0.5.0
Django==2.2.10
django-background-tasks==1.2.5
django-compat==1.0.15
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
